fix: resolver 53 findings da revisão de código do FIX.md
- O15: escopo de organização no RebalanceService para multi-tenancy
- O4: ALLOWED_REDIRECT_HOSTS configurável via SecuritySettings
- O5: evolução TTL dos estados OIDC (10 min) com tratamento seguro
- O7: parsing host:port seguro para IPv6 no readiness.py
- O17: detecção robusta de 'already exists' no investment_candidates
- R5-16/17: deletar shim agents.py, atualizar 13 imports para audit_models
- R5-19: adicionar updated_at em StrategyMandate, onupdate em TradeIntent
- R5-24: sa.JSON → JSONB em MetricDefinition.dependencies
- R5-25: Float → Numeric para Position.weight_pct, PortfolioConstraint.limit_value
- R5-26: CHECK (original_id <> duplicate_id) em EventDuplicate
- R5-31: HNSW m=16 ef_construction=200 em DocumentChunk, novo HNSW em Embedding
- R5-32: lazy='selectin' → lazy='select' em todos relationships de InvestmentCandidateRecord
- R5: Float → Numeric em news.py (5), processing.py (2), audit.py (1)
- R9-13: permissões CI: read-all em ci.yml, frontend.yml
- R9-21: sintaxe bash no AGENTS.md (PowerShell removido)
- R9-23: versão do Node padronizada para 22
- R8: 30 novos testes — rebalance service (18) + rebalance routes (12)
- Arquivo: UPGRADE-1.md, PLAN-3.md, FIX.md → docs/plan/v3/
- Novo: docs/plan/v3/pendente.md para itens não-MVP

1271 passando, 0 falhas

+++ src/database/models/audit.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import json
 from datetime import datetime
+from decimal import Decimal
 from hashlib import sha256
 from typing import Any
 from uuid import UUID
@@ -49,15 +50,13 @@ class AuditLogEntry(Base):
     request_id: Mapped[UUID | None] = mapped_column(PG_UUID(as_uuid=True), nullable=True, index=True)
     http_method: Mapped[str | None] = mapped_column(sa.String(7), nullable=True)
     http_path: Mapped[str | None] = mapped_column(sa.String(500), nullable=True)
-    duration_ms: Mapped[float | None] = mapped_column(sa.Float, nullable=True)
+    duration_ms: Mapped[Decimal | None] = mapped_column(sa.Numeric(12, 3), nullable=True)
     status_code: Mapped[int | None] = mapped_column(sa.Integer, nullable=True)
     hash_prev: Mapped[str | None] = mapped_column(sa.String(64), nullable=True)
     hash: Mapped[str] = mapped_column(sa.String(64), nullable=False, unique=True)
     timestamp: Mapped[datetime] = mapped_column(sa.DateTime(timezone=True), nullable=False, default=utcnow)
     created_at: Mapped[datetime] = mapped_column(sa.DateTime(timezone=True), nullable=False, default=utcnow)
-    updated_at: Mapped[datetime] = mapped_column(
-        sa.DateTime(timezone=True), default=utcnow, onupdate=utcnow
-    )
+    updated_at: Mapped[datetime] = mapped_column(sa.DateTime(timezone=True), default=utcnow, onupdate=utcnow)
 
     __table_args__ = (
         sa.Index("ix_audit_log_tenant_timestamp", "tenant_id", "timestamp"),
